+++ tools/registry.js
@@ -12,11 +12,13 @@ const getEnvVars = require('./get-env-vars');
     parser.add_argument('--input', { help: 'Input file path, for importing to Roku device' });
     parser.add_argument('--output', { help: 'Output file path, for exporting from Roku device' });
     parser.add_argument('--clear', { help: 'Clear registry', action: BooleanOptionalAction, default: false });
+    parser.add_argument('--raw', { help: 'Keep section values as is', action: BooleanOptionalAction, default: false });
 
     const args = parser.parse_args();
     const input = args.input;
     const output = args.output;
     const clear = args.clear;
+    const raw = args.raw;
 
     // only one action can be specified at a time between input, output and clear
     const actions = [input, output, clear].filter(Boolean);
@@ -32,9 +34,11 @@ const getEnvVars = require('./get-env-vars');
     if (input) {
         const registry = JSON.parse(fs.readFileSync(input, { encoding: 'utf8', flag: 'r' }));
 
-        for (const section in registry) {
-            for (const key in registry[section]) {
-                registry[section][key] = JSON.stringify(registry[section][key]);
+        if (!raw) {
+            for (const section in registry) {
+                for (const key in registry[section]) {
+                    registry[section][key] = JSON.stringify(registry[section][key]);
+                }
             }
         }
 
@@ -62,12 +66,14 @@ const getEnvVars = require('./get-env-vars');
 
         const registry = await response.json();
 
-        for (const section in registry) {
-            for (const key in registry[section]) {
-                try {
-                    registry[section][key] = JSON.parse(registry[section][key]);
-                } catch (error) {
-                    console.error(`Failed to parse value for key "${key}" in section "${section}"`);
+        if (!raw) {
+            for (const section in registry) {
+                for (const key in registry[section]) {
+                    try {
+                        registry[section][key] = JSON.parse(registry[section][key]);
+                    } catch (error) {
+                        console.error(`Failed to parse value for key "${key}" in section "${section}"`);
+                    }
                 }
             }
         }